MessengerApp
Developed an instant messaging android application that enables a user
to chat with a friend. The user can login using their Google or Facebook
credentials, edit and update his/her profile content and can chat with
all the user listings
in the application. Android Studio and a real-time database Google
Firebase were used in developing this application.